Move `Checkout` component to checkout subfolder
