+++ src/components/checkout/Checkout.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-import CheckoutMessage from './checkout/CheckoutMessage.svelte'
+import CheckoutMessage from './CheckoutMessage.svelte'
 import type { PolicyItem } from 'data/items'
 import { selectedPolicy } from 'data/policies'
 import { HOME, TERMS_OF_SERVICE } from 'helpers/routes'
-import ItemDeleteModal from './ItemDeleteModal.svelte'
-import ItemDetails from './ItemDetails.svelte'
+import ItemDeleteModal from '../ItemDeleteModal.svelte'
+import ItemDetails from '../ItemDetails.svelte'
 import { goto } from '@roxi/routify'
 import { Button, Checkbox } from '@silintl/ui-components'
 import { createEventDispatcher } from 'svelte'
